// src/app/core/model/student-model.ts
export interface IStudent {
  studentId: number;
  refrenceById: number;
  instituteId: number;
  name: string;
  contactNo: string;
  email: string;
  city: string;
  state: string;
  pincode: string;
  qualification: string;
  collegeName: string;
  collegeCity: string;
  familyDetails: string;
  aadharCard: string;
  profilePhotoName: string;
}